Move some utils into seperate file
