+++ json_codegen/generators/python3.py
@@ -4,7 +4,7 @@ from re import search
 from json_codegen.ast import python as ast
 from json_codegen.core import SchemaParser, BaseGenerator
 from json_codegen.generators.python3object import Python3ObjectGenerator
-from . import marshmallow_type_map
+from json_codegen.python_utils import class_name, marshmallow_type_map, upper_first_letter
 
 
 class Python3Generator(SchemaParser, BaseGenerator):
@@ -69,7 +69,7 @@ class Python3Generator(SchemaParser, BaseGenerator):
 
         # Create class definition
         class_def = ast.ClassDef(
-            name=Python3Generator.upper_first_letter(definition["title"]) + "Schema",
+            name=upper_first_letter(definition["title"]) + "Schema",
             bases=[ast.Name(id="Schema")],
             body=class_body,
             decorator_list=[],
@@ -193,7 +193,7 @@ class Python3Generator(SchemaParser, BaseGenerator):
 
         # Where the array type references a definition, make a nested field with
         # the type of the item schema
-        ref_title = Python3Generator.upper_first_letter(ref["title"])
+        ref_title = upper_first_letter(ref["title"])
         for node in ast.walk(value):
             if isinstance(node, ast.Call):
                 x = self._make_field("Nested", [ast.Name(id=ref_title + "Schema")], [])
@@ -241,7 +241,7 @@ class Python3Generator(SchemaParser, BaseGenerator):
 
     @staticmethod
     def construct_dec_post_load(schema):
-        name = Python3ObjectGenerator.class_name(schema.name)
+        name = class_name(schema.name)
         name_lower = name.lower()
 
         fn_args = ast.arguments(
@@ -267,14 +267,6 @@ class Python3Generator(SchemaParser, BaseGenerator):
             decorator_list=[ast.Name(id="post_load")],
             returns=None,
         )
-
-    @staticmethod
-    def upper_first_letter(s):
-        """
-        Assumes custom types of two words are defined as customType
-        such that the class name is CustomTypeSchema
-        """
-        return s[0].upper() + s[1:]
 
     def as_ast(self):
         return ast.Module(body=self._body)
